Made buddy header one word, so returned address is aligned
Now everythin works as is, up to -O3

+++ buddy.h
@@ -4,7 +4,7 @@
 #define KERNEL_MAX_ORDER   15
 #define KERNEL_MIN_ORDER   4
 
-#define BUDDY_HEADER_SIZE   sizeof(uint8_t)
+#define BUDDY_HEADER_SIZE   sizeof(uint32_t)
 
 struct heapnode {
     uint8_t order;
